chore: remove todo and 'expiry' arg from PaymentFlowStateRepository
Staying with the long-persistence in mongo instead of expiring records
in redis because there's a path where we'd need previously constructed
flows to settle pending payments.

We can optionally add code to delete payment flows thru fire-and-forget
calls or a task in the cronjob for completed payments (successful or
failed) to address the issue of keeping around unnecessary data.